cond input made categorical | embedding instead of linear proj. Still not enough for overfit

+++ mint/core/inputs.py
@@ -62,7 +62,7 @@ def create_input(train_eval_config,
   # conditioning
   name_to_features.update({
     'conditioning_name': tf.io.FixedLenFeature([], tf.string),
-    'conditioning_input': tf.io.VarLenFeature(tf.float32),
+    'conditioning_input': tf.io.VarLenFeature(tf.int64),
   })
 
   # For training, we want a lot of parallel reading.
